Close #430 - Rename package - effectie-monix3: effectie.monix => effectie.monix3
