+++ modules/effectie-monix3/shared/src/main/scala/effectie/monix3/fromFuture.scala
@@ -1,4 +1,4 @@
-package effectie.monix
+package effectie.monix3
 
 import cats.Id
 import cats.effect.{ContextShift, IO}
