+++ modules/effectie-monix3/shared/src/main/scala/effectie/monix3/fxCtor.scala
@@ -1,4 +1,4 @@
-package effectie.monix
+package effectie.monix3
 
 import effectie.core.FxCtor
 import monix.eval.Task
