+++ modules/effectie-monix3/shared/src/test/scala/effectie/monix3/MonadSpec.scala
@@ -1,4 +1,4 @@
-package effectie.monix
+package effectie.monix3
 
 import cats.{Eq, Monad}
 import hedgehog.runner.Test
