+++ modules/effectie-monix3/shared/src/test/scala/effectie/monix3/canCatchSpec.scala
@@ -1,5 +1,6 @@
-package effectie.monix
+package effectie.monix3
 
+import canCatch._
 import cats._
 import cats.data.EitherT
 import cats.effect.IO
@@ -7,12 +8,12 @@ import cats.instances.all._
 import cats.syntax.all._
 import effectie.SomeControlThrowable
 import effectie.core._
-import effectie.monix.canCatch._
-import effectie.monix.fxCtor._
+import effectie.syntax.error._
 import effectie.syntax.fx._
 import effectie.testing.types.SomeError
 import extras.concurrent.testing.ConcurrentSupport
 import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
+import fxCtor._
 import hedgehog._
 import hedgehog.runner._
 import monix.eval.Task
